requirements-gpu.txt
matplotlib==3.2.2
numpy==1.19.2
Pillow==8.2.0
pytorch==1.3.1
torchvision==0.4.2
tensorboard==2.4.0
future==0.18.2
tqdm==4.59.0
scikit-image==0.17.2
tensorboardX
ml-collections
medpy
SimpleITK
scipy
h5py
timm
einops

